Add multi_encaps that can encrypt n recipient known at run time
- We also move the test macros to a seperate file/module.
- We disabled some versions of cgwkv1-3 multi-encaps as this is probably
  not the best way to implement multi-encaps for these schemes.
