+++ src/kem/cgw_kv3.rs
@@ -7,7 +7,7 @@
 //!
 //! Important notice: Keep in mind that the security of this scheme has not formally been proven.
 
-use crate::kem::{DecapsulationError, SharedSecret, IBKEM};
+use crate::kem::{Error, SharedSecret, IBKEM};
 use crate::util::*;
 use crate::Compress;
 use core::convert::TryInto;
@@ -212,8 +212,34 @@ impl IBKEM for CGWKV3 {
         id: &Identity,
         rng: &mut R,
     ) -> (CipherText, SharedSecret) {
-        let (cts, k) = Self::multi_encaps::<R, 1>(pk, &[id], rng);
-        (cts[0], k)
+        let k = rand_gt(rng);
+
+        let s = rand_scalar(rng);
+        let x = id.to_scalar();
+        let pre_y: G1Affine = (pk.a_1[0] * s).into();
+        let y = hash_g1_to_scalar(pre_y);
+
+        let cprime = pk.kta_t * s + k;
+
+        let batch = [
+            (pk.a_1[1] * s),
+            (pk.w0ta_1[0] * s) + (pk.w1ta_1[0] * (s * x)),
+            (pk.w0ta_1[1] * s) + (pk.w1ta_1[1] * (s * x)),
+            (pk.wprime_1[0] * s) + (pk.wprime_1[1]) * (s * y),
+        ];
+
+        let mut out = [G1Affine::default(); 4];
+        G1Projective::batch_normalize(&batch, &mut out);
+
+        (
+            CipherText {
+                c0: [pre_y, out[0]],  // C_i
+                c1: [out[1], out[2]], // C'_i
+                c2: out[3],           // C''
+                cprime,
+            },
+            SharedSecret::from(&k),
+        )
     }
 
     /// Derive the same SharedSecret from the CipherText using a UserSecretKey.
@@ -223,7 +249,7 @@ impl IBKEM for CGWKV3 {
         _ompk: Option<&PublicKey>,
         usk: &UserSecretKey,
         ct: &CipherText,
-    ) -> Result<SharedSecret, DecapsulationError> {
+    ) -> Result<SharedSecret, Error> {
         let y = hash_g1_to_scalar(ct.c0[0]);
         let z: G2Affine = (usk.d2[0] + (usk.d2[1] * y)).into();
 
@@ -246,45 +272,45 @@ impl IBKEM for CGWKV3 {
     }
 }
 
-impl CGWKV3 {
-    /// Generate a SharedSecret and corresponding Ciphertext for that key.
-    pub fn multi_encaps<R: Rng + CryptoRng, const N: usize>(
-        pk: &PublicKey,
-        ids: &[&Identity; N],
-        rng: &mut R,
-    ) -> ([CipherText; N], SharedSecret) {
-        let k = rand_gt(rng);
-
-        let mut cts = [CipherText::default(); N];
-        for (i, id) in ids.iter().enumerate() {
-            let s = rand_scalar(rng);
-            let x = id.to_scalar();
-            let pre_y: G1Affine = (pk.a_1[0] * s).into();
-            let y = hash_g1_to_scalar(pre_y);
-
-            let cprime = pk.kta_t * s + k;
-
-            let batch = [
-                (pk.a_1[1] * s),
-                (pk.w0ta_1[0] * s) + (pk.w1ta_1[0] * (s * x)),
-                (pk.w0ta_1[1] * s) + (pk.w1ta_1[1] * (s * x)),
-                (pk.wprime_1[0] * s) + (pk.wprime_1[1]) * (s * y),
-            ];
-
-            let mut out = [G1Affine::default(); 4];
-            G1Projective::batch_normalize(&batch, &mut out);
-
-            cts[i] = CipherText {
-                c0: [pre_y, out[0]],  // C_i
-                c1: [out[1], out[2]], // C'_i
-                c2: out[3],           // C''
-                cprime,
-            }
-        }
-
-        (cts, SharedSecret::from(&k))
-    }
-}
+//impl CGWKV3 {
+//    /// Generate a SharedSecret and corresponding Ciphertext for that key.
+//    pub fn multi_encaps<R: Rng + CryptoRng, const N: usize>(
+//        pk: &PublicKey,
+//        ids: &[&Identity; N],
+//        rng: &mut R,
+//    ) -> ([CipherText; N], SharedSecret) {
+//        let k = rand_gt(rng);
+//
+//        let mut cts = [CipherText::default(); N];
+//        for (i, id) in ids.iter().enumerate() {
+//            let s = rand_scalar(rng);
+//            let x = id.to_scalar();
+//            let pre_y: G1Affine = (pk.a_1[0] * s).into();
+//            let y = hash_g1_to_scalar(pre_y);
+//
+//            let cprime = pk.kta_t * s + k;
+//
+//            let batch = [
+//                (pk.a_1[1] * s),
+//                (pk.w0ta_1[0] * s) + (pk.w1ta_1[0] * (s * x)),
+//                (pk.w0ta_1[1] * s) + (pk.w1ta_1[1] * (s * x)),
+//                (pk.wprime_1[0] * s) + (pk.wprime_1[1]) * (s * y),
+//            ];
+//
+//            let mut out = [G1Affine::default(); 4];
+//            G1Projective::batch_normalize(&batch, &mut out);
+//
+//            cts[i] = CipherText {
+//                c0: [pre_y, out[0]],  // C_i
+//                c1: [out[1], out[2]], // C'_i
+//                c2: out[3],           // C''
+//                cprime,
+//            }
+//        }
+//
+//        (cts, SharedSecret::from(&k))
+//    }
+//}
 
 impl Compress for PublicKey {
     const OUTPUT_SIZE: usize = PK_BYTES;
@@ -527,5 +553,5 @@ mod tests {
     use crate::Derive;
 
     test_kem!(CGWKV3);
-    test_multi_kem!(CGWKV3);
+    //test_multi_kem!(CGWKV3);
 }
